lab 2 perfect
