--- tex_prog/tex_prog.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31624.102
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "tex_prog", "tex_prog\tex_prog.csproj", "{30B5153F-DC9E-4B9C-952E-D730136C8391}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{30B5153F-DC9E-4B9C-952E-D730136C8391}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{30B5153F-DC9E-4B9C-952E-D730136C8391}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{30B5153F-DC9E-4B9C-952E-D730136C8391}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{30B5153F-DC9E-4B9C-952E-D730136C8391}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {74DC8BAE-A5F5-4FCF-9FD4-7F77F7FC4760}
-	EndGlobalSection
-EndGlobal
